Added Order Funcions

+++ src/lk/ijse/dep/app/dao/custom/OrderDAO.java
@@ -5,6 +5,6 @@ import lk.ijse.dep.app.entity.Order;
 
 public interface OrderDAO extends CrudDAO<Order, String> {
 
-    int count() throws Exception;
+    Long count() throws Exception;
 
 }
